requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-docx==1.1.2
requests==2.32.3
openai==1.30.1
anthropic==0.26.1
google-generativeai==0.7.2
python-dotenv==1.0.1
reportlab==4.2.2
